Refactor `build.zig` to allow manual creation of executables
The previous method of adding executables to our builder using a
helper function prevented us from being able to modify the build steps
of said executable in a neat manner. For this reason the helper function
was removed.

A new helper function `addRunStep` was added in its place to easily add
a run step to a given executable.